+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ##### New Features
 
-* **plug:** add metrix plug for response time metrics ([#3](https://github.com/mpiercy827/metrix/pull/3))
-* **function-timer:** add FunctionTimer module ([#2](https://github.com/mpiercy827/metrix/pull/2))
+* **plug:** add ex_metrics plug for response time metrics ([#3](https://github.com/mpiercy827/ex_metrics/pull/3))
+* **function-timer:** add FunctionTimer module ([#2](https://github.com/mpiercy827/ex_metrics/pull/2))
 
 ##### Chores
 
-* **base:** add project base ([#1](https://github.com/mpiercy827/metrix/pull/2))
+* **base:** add project base ([#1](https://github.com/mpiercy827/ex_metrics/pull/2))